+++ src/rigidbody.hpp
@@ -6,9 +6,9 @@ class Rigidbody : public Entity {
     Vector3 velocity;
 
    public:
-    unsigned int mass;
+    float mass;
 
-    Rigidbody(Model& model, Vector3 position, unsigned int mass);
+    Rigidbody(Model& model, Vector3 position, float mass);
 
     void update();
     void applyForce(Vector3 force);
